backend/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
gunicorn>=21.0.0

# Database & Security
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
alembic>=1.11.0

# Image handling
Pillow>=9.0.0
cloudinary==1.41.0

# Core AI (Versions pinned for stability)
tensorflow-cpu==2.15.0
numpy==1.26.4
scikit-learn>=1.0.0

# Image Processing & Grad-CAM
opencv-python-headless>=4.5.0
matplotlib>=3.5.0

# Analysis & Notebooks
pandas>=1.3.0
seaborn>=0.11.0
tqdm>=4.62.0
jupyter>=1.0.0
ipykernel>=6.0.0